progress in fixing pypy issues

+++ joblib/_memmapping_reducer.py
@@ -575,14 +575,17 @@ class TemporaryResourcesManager(object):
             for context_id in list(self._cached_temp_folders):
                 self._unregister_context(context_id)
         else:
-            temp_folder = self._cached_temp_folders[context_id]
-            finalizer = self._finalizers[context_id]
+            temp_folder = self._cached_temp_folders.get(context_id)
+            finalizer = self._finalizers.get(context_id)
 
-            resource_tracker.unregister(temp_folder, "folder")
-            atexit.unregister(finalizer)
+            if temp_folder is not None:
+                resource_tracker.unregister(temp_folder, "folder")
 
-            self._cached_temp_folders.pop(context_id)
-            self._finalizers.pop(context_id)
+            if finalizer is not None:
+                atexit.unregister(finalizer)
+
+            self._cached_temp_folders.pop(context_id, None)
+            self._finalizers.pop(context_id, None)
 
     # resource management API
 
@@ -623,7 +626,9 @@ class TemporaryResourcesManager(object):
             for context_id in list(self._cached_temp_folders):
                 self._unlink_temporary_resources(context_id)
         else:
-            temp_folder = self._cached_temp_folders[context_id]
+            temp_folder = self._cached_temp_folders.get(context_id)
+            if temp_folder is None:
+                return
             if os.path.exists(temp_folder):
                 for filename in os.listdir(temp_folder):
                     resource_tracker.maybe_unlink(
@@ -641,7 +646,9 @@ class TemporaryResourcesManager(object):
             for context_id in list(self._cached_temp_folders):
                 self._unregister_temporary_resources(context_id)
         else:
-            temp_folder = self._cached_temp_folders[context_id]
+            temp_folder = self._cached_temp_folders.get(context_id)
+            if temp_folder is None:
+                return
             if os.path.exists(temp_folder):
                 for filename in os.listdir(temp_folder):
                     resource_tracker.unregister(
@@ -657,7 +664,9 @@ class TemporaryResourcesManager(object):
                     allow_non_empty=allow_non_empty, context_id=context_id
                 )
         else:
-            temp_folder = self._cached_temp_folders[context_id]
+            temp_folder = self._cached_temp_folders.get(context_id)
+            if temp_folder is None:
+                return
             try:
                 delete_folder(
                     temp_folder, allow_non_empty=allow_non_empty
